+++ lib/godmin/resources/resource_controller.rb
@@ -154,9 +154,17 @@ module Godmin
         return false unless params[:batch_action].present?
 
         item_ids = params[:id].split(",").map(&:to_i)
+        records = @resource_class.find(item_ids)
+        if authorization_enabled?
+          records = records.select { |r| policy(r).send("batch_action_#{params[:batch_action]}?") }
+        end
 
-        if @resource_service.batch_action(params[:batch_action], item_ids)
+        if @resource_service.batch_action(params[:batch_action], records)
           flash[:updated_ids] = item_ids
+          flash[:notice] = translate_scoped("flash.batch_action",
+                                            number_of_affected_records: records.length,
+                                            total_number_of_records: item_ids.length,
+                                            resource: @resource_class.model_name.human(count: records.length))
 
           if respond_to?("redirect_after_batch_action_#{params[:batch_action]}", true)
             redirect_to send("redirect_after_batch_action_#{params[:batch_action]}") and return true